requirements.txt
aiogram
qrcode
python-dotenv
Pillow
